moved files for flyingship to subdir "spells"
