+++ src/move.c
@@ -28,7 +28,6 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include "monster.h"
 #include "lighthouse.h"
 #include "piracy.h"
-#include "flyingship.h"
 
 #include <kernel/build.h>
 #include <kernel/building.h>
@@ -48,6 +47,8 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <kernel/terrain.h>
 #include <kernel/terrainid.h>
 #include <kernel/unit.h>
+
+#include <spells/flyingship.h>
 
 #include "direction.h"
 #include "calendar.h"
